fix: conversion overflows

+++ src/libraries/Converter.sol
@@ -9,7 +9,7 @@ library Converter {
     function _amountFromSats(uint64 amountSats) internal pure returns (uint256) {
         uint8 TokenDecimals = Constants.TokenDecimals;
         if (TokenDecimals >= BtcDecimals) {
-            return uint256(amountSats * uint64(10 ** (TokenDecimals - BtcDecimals)));
+            return uint256(amountSats) * 10 ** (TokenDecimals - BtcDecimals);
         } else {
             return uint256(amountSats / uint64(10 ** (BtcDecimals - TokenDecimals)));
         }
